+++ v-next/hardhat-utils/test/synchronization.ts
@@ -2,7 +2,7 @@ import assert from "node:assert/strict";
 import { describe, it } from "node:test";
 
 import { sleep } from "../src/lang.js";
-import { MultiProcessMutex } from "../src/synchronization.js";
+import { AsyncMutex, MultiProcessMutex } from "../src/synchronization.js";
 
 describe("multi-process-mutex", () => {
   const mutexName = "test-mutex";
@@ -99,5 +99,30 @@ describe("multi-process-mutex", () => {
       duration < 1000,
       "Duration should be less than the mutex timeout",
     );
+  });
+});
+
+describe("AsyncMutex", () => {
+  it("should run a function exclusively", async () => {
+    const mutex = new AsyncMutex();
+
+    let running = 0;
+    let maxRunning = 0;
+
+    async function run() {
+      running++;
+      maxRunning = Math.max(maxRunning, running);
+      await new Promise((resolve) => setTimeout(resolve, 10));
+      running--;
+    }
+
+    await Promise.all([
+      mutex.exclusiveRun(run),
+      mutex.exclusiveRun(run),
+      mutex.exclusiveRun(run),
+    ]);
+
+    assert.equal(maxRunning, 1);
+    assert.equal(running, 0);
   });
 });
