+++ v-next/hardhat/src/internal/builtin-plugins/network-manager/hook-handlers/network.ts
@@ -10,8 +10,8 @@ import type {
 import type { RequestHandler } from "../request-handlers/types.js";
 
 import { deepClone } from "@nomicfoundation/hardhat-utils/lang";
+import { AsyncMutex } from "@nomicfoundation/hardhat-utils/synchronization";
 
-import { AsyncMutex } from "../../../core/async-mutex.js";
 import { isJsonRpcResponse } from "../json-rpc.js";
 
 export default async (): Promise<Partial<NetworkHooks>> => {
